refactor: move files to /constants directory
Those files fit the description of the directory

refactor: move declarations files to a types dir

A lot cleaner to separate them from the rest of the app

refactor: relocate <App> and related files

Only the index.js should be in the root directory

refactor: relocate providers in components

A provider is a component, and is not a screen, so it has to go in components

refactor: move Sentry.js to /libs

It had no businees being in the root folder like other libs

refactor: move colors.js file to ui/

The theme folder had only one file, better to regroup ui files.

fix: update all pathnames to import recently moved files

fix: remove old dev.ts config

it is now located in constants/dev-config.ts
fix: keep app files in src root

fix: keep ignorelogbox to false
