+++ src/App.js
@@ -11,6 +11,7 @@ import { CozyProvider, useClient } from 'cozy-client'
 import { NativeIntentProvider } from 'cozy-intent'
 
 import * as RootNavigation from './libs/RootNavigation'
+import NetStatusBoundary from './libs/services/NetStatusBoundary'
 import { CozyAppScreen } from './screens/cozy-app/CozyAppScreen'
 import { CreateInstanceScreen } from './screens/login/CreateInstanceScreen'
 import { CryptoWebView } from './components/webviews/CryptoWebView/CryptoWebView'
@@ -20,18 +21,17 @@ import { HttpServerProvider } from './libs/httpserver/httpServerProvider'
 import { LockScreen } from '/screens/lock/LockScreen'
 import { LoginScreen } from './screens/login/LoginScreen'
 import { OnboardingScreen } from './screens/login/OnboardingScreen'
-import { SplashScreenProvider } from './providers/SplashScreenProvider'
+import { SplashScreenProvider } from '/components/providers/SplashScreenProvider'
 import { WelcomeScreen } from './screens/welcome/WelcomeScreen'
 import { getClient } from './libs/client'
-import { getColors } from './theme/colors'
+import { getColors } from '/ui/colors'
 import { localMethods } from './libs/intents/localMethods'
-import { routes } from './constants/routes.js'
+import { routes } from '/constants/routes.js'
 import { useAppBootstrap } from './hooks/useAppBootstrap.js'
-import { useNetService } from '/libs/services/NetService'
-import { withSentry } from './Sentry'
 import { useCookieResyncOnResume } from './hooks/useCookieResyncOnResume'
 import { useGlobalAppState } from './hooks/useGlobalAppState'
-import NetStatusBoundary from './libs/services/NetStatusBoundary'
+import { useNetService } from '/libs/services/NetService'
+import { withSentry } from '/libs/Sentry'
 
 const Root = createStackNavigator()
 const Stack = createStackNavigator()
