+++ src/components/providers/SplashScreenProvider.tsx
