+++ src/hooks/useAppBootstrap.js
@@ -2,8 +2,8 @@ import Minilog from '@cozy/minilog'
 import { Linking, LogBox } from 'react-native'
 import { useEffect, useState } from 'react'
 
-import { SentryTags, setSentryTag } from '/Sentry'
-import { devConfig } from '/config/dev'
+import { SentryTags, setSentryTag } from '/libs/Sentry'
+import { devConfig } from '/constants/dev-config'
 import { manageIconCache } from '/libs/functions/iconTable'
 import { navigate } from '/libs/RootNavigation'
 import { routes } from '/constants/routes'
