+++ src/libs/services/NetService.ts
@@ -7,8 +7,8 @@ import { useEffect } from 'react'
 import CozyClient from 'cozy-client'
 import Minilog from '@cozy/minilog'
 
-import strings from '/strings.json'
-import { devConfig } from '/config/dev'
+import strings from '/constants/strings.json'
+import { devConfig } from '/constants/dev-config'
 import { reset } from '/libs/RootNavigation'
 import { routes } from '/constants/routes'
 import { showSplashScreen } from '/libs/services/SplashScreenService'
