+++ src/screens/login/CreateInstanceScreen.js
@@ -8,7 +8,7 @@ import { routes } from '/constants/routes'
 import { useDimensions } from '/libs/dimensions'
 import { consumeRouteParameter } from '/libs/functions/routeHelpers'
 
-import { getColors } from '/theme/colors'
+import { getColors } from '/ui/colors'
 
 import { ClouderyCreateInstanceView } from './components/ClouderyCreateInstanceView'
