+++ src/screens/login/LoginScreen.js
@@ -24,8 +24,8 @@ import { resetKeychainAndSaveLoginData } from '/libs/functions/passwordHelpers'
 import { consumeRouteParameter } from '/libs/functions/routeHelpers'
 import { useSplashScreen } from '/hooks/useSplashScreen'
 
-import strings from '../../strings.json'
-import { getColors } from '/theme/colors'
+import strings from '/constants/strings.json'
+import { getColors } from '/ui/colors'
 import { NetService } from '/libs/services/NetService'
 import { routes } from '/constants/routes'
